wip: move result mapper to interface/http
