+++ src/framework/normalizer/normalize.ts
@@ -1,0 +1,4 @@
+import { type Normalizer } from '@/interface/http/normalizer.type';
+import { createNormalizer } from '@koala-ts/framework';
+
+export const normalize: Normalizer = createNormalizer([]);
