+++ src/interface/http/map-result-to-http.ts
@@ -1,0 +1,4 @@
+import { normalize } from '@/framework/normalizer/normalize';
+import { createResultToHttpMapper } from '@/interface/http/result-to-http';
+
+export const mapResultToHttp = createResultToHttpMapper({ normalize: normalize });
