+++ src/interface/http/result-to-http.ts
@@ -1,3 +1,4 @@
+import { type Normalizer, type SerializeOptions } from '@/interface/http/normalizer.type';
 import { HTTP_INTERNAL_SERVER_ERROR, type StatusCode } from '@/kairos/shared/http/status-code';
 import { isErr, type Result } from '@/kairos/shared/result/result';
 
@@ -7,19 +8,6 @@ export interface HttpResponsePayload {
 }
 
 type StatusResolver<T> = StatusCode | ((value: T) => StatusCode);
-interface PropertyMetadata {
-  ignore?: boolean;
-  groups?: string[];
-  serializedName?: string;
-  metadata?: Metadata;
-}
-
-type Metadata = Record<string, PropertyMetadata>;
-
-interface SerializeOptions {
-  groups?: string[];
-  metadata?: Metadata;
-}
 type SerializeOptionsResolver<T> = SerializeOptions | ((value: T) => SerializeOptions | undefined);
 
 interface ErrorHttpResponse<E> {
@@ -42,10 +30,10 @@ export interface ResultHttpMapping<T, E extends { type: string }> {
 }
 
 type CreateResultToHttpMapperDependencies = Readonly<{
-  normalize: (input: unknown, options?: SerializeOptions) => unknown;
+  normalize: Normalizer;
 }>;
 
-type ResultToHttpMapper = <T, E extends { type: string }>(
+export type ResultToHttpMapper = <T, E extends { type: string }>(
   result: Result<T, E>,
   mapping: ResultHttpMapping<T, E>,
 ) => HttpResponsePayload;
