+++ src/kairos/authentication/interface/http/login-handler.ts
@@ -1,7 +1,7 @@
+import { mapResultToHttp } from '@/interface/http/map-result-to-http';
 import { login } from '@/kairos/authentication/authentication-composition';
 import { type LoginRequest } from '@/kairos/authentication/interface/http/login-request';
 import { loginResponse } from '@/kairos/authentication/interface/http/login-response';
-import { mapResultToHttp } from '@/kairos/shared/http/map-result-to-http';
 import { type HttpScope } from '@koala-ts/framework';
 
 export async function loginHandler({ response, request }: HttpScope): Promise<void> {
