+++ src/kairos/workspace/interface/http/create-workspace-response.ts
@@ -1,4 +1,4 @@
-import { type ResultHttpMapping } from '@/kairos/shared/http/result-to-http';
+import { type ResultHttpMapping } from '@/interface/http/result-to-http';
 import { HTTP_CONFLICT, HTTP_CREATED } from '@/kairos/shared/http/status-code';
 import type { WorkspaceSlugConflictError } from '@/kairos/workspace/domain/errors';
 import type { Workspace } from '@/kairos/workspace/domain/workspace';
